+++ starkinfra/pixrequest/log/__log.py
@@ -48,7 +48,7 @@ def query(limit=None, after=None, before=None, types=None, request_ids=None, use
     - limit [integer, default None]: maximum number of objects to be retrieved. Unlimited if None. ex: 35
     - after [datetime.date or string, default None]: date filter for objects created or updated only after specified date. ex: datetime.date(2020, 3, 10)
     - before [datetime.date or string, default None]: date filter for objects created or updated only before specified date. ex: datetime.date(2020, 3, 10)
-    - types [string, default None]: filter retrieved objects by types. ex: "success" or "failed"
+    - types [string, default None]: filter retrieved objects by types. ex: ["success", "failed"]
     - request_ids [list of strings, default None]: list of PixRequest IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
     - user [Organization/Project object, default None]: Organization or Project object. Not necessary if starkinfra.user was set before function call
     ## Return:
@@ -74,7 +74,7 @@ def page(cursor=None, limit=None, after=None, before=None, types=None, request_i
     - limit [integer, default None]: maximum number of objects to be retrieved. Unlimited if None. ex: 35
     - after [datetime.date or string, default None]: date filter for objects created or updated only after specified date. ex: datetime.date(2020, 3, 10)
     - before [datetime.date or string, default None]: date filter for objects created or updated only before specified date. ex: datetime.date(2020, 3, 10)
-    - types [string, default None]: filter retrieved objects by types. ex: "success" or "failed"
+    - types [string, default None]: filter retrieved objects by types. ex: ["success", "failed"]
     - request_ids [list of strings, default None]: list of PixRequest IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
     - user [Organization/Project object, default None]: Organization or Project object. Not necessary if starkinfra.user was set before function call
     ## Return: