+++ starkinfra/utils/__init__.py
@@ -1,0 +1,1 @@
+from starkcore.utils.cache import cache