+++ CHANGELOG.md
@@ -15,6 +15,15 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - CreditNote resource for money lending with Stark Infra's endorsement
+- IssuingAuthorization resource for Sub Issuers
+- IssuingBalance resource for Sub Issuers
+- IssuingBin resource for Sub Issuers
+- IssuingCard resource for Sub Issuers
+- IssuingHolder resource for Sub Issuers
+- IssuingInvoice resource for Sub Issuers
+- IssuingPurchase resource for Sub Issuers
+- IssuingTransaction resource for Sub Issuers
+- IssuingWithdrawal resource for Sub Issuers
 
 ## [0.0.2] - 2022-03-15
 ### Added